added controller for user data endpoints
